Big scheduler rewrite, fix page allocator bug
* The scheduler now works with threads, this allows one process to have
  more than one thread of execution
* The scheduler is a lot nicer to work with now, still a lot of work to
  be done
* Fix a nasty bug in the ide driver that caused a kernel stack overflow
  and corrupted the page frame allocator's bitmap. This caused it to
  allocate already allocated pages :P
* A bunch of smaller stuff here and there

+++ libs/data/bitmap.h
@@ -6,7 +6,7 @@
 
 #define BITMAP_WORD_SIZE (sizeof(bitmap_word) * CHAR_BIT)
 
-typedef u64 bitmap_word;
+typedef u32 bitmap_word;
 
 
 void bitmap_set(bitmap_word* bitmap, usize index);
